[31] - Controller responsability separated
