+++ app/Application/Exceptions/WalletNotFoundException.php
@@ -2,13 +2,13 @@
 
 namespace App\Application\Exceptions;
 
-use Exception;
 use Symfony\Component\HttpFoundation\Response;
+use Exception;
 
-class UserNotFoundException extends Exception
+class WalletNotFoundException extends Exception
 {
     public function __construct()
     {
-        parent::__construct('User not found', Response::HTTP_BAD_REQUEST);
+        parent::__construct('Wallet not found', Response::HTTP_BAD_REQUEST);
     }
 }
